# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/leagues/{guildId}/season:
    get:
      operationId: getActiveSeason
      tags: [league]
      summary: Get active season
      description: Returns the currently active season for a guild including all config overrides
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: Active season
          content:
            application/json:
              schema:
                type: object
                properties:
                  season:
                    $ref: "#/components/schemas/Season"
                required: [season]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/teams:
    get:
      operationId: getLeagueTeams
      tags: [league]
      summary: Get all franchise teams
      description: Returns all MCA teams for the active season including human/CPU status and Discord IDs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  teams:
                    type: array
                    items:
                      $ref: "#/components/schemas/McaTeam"
                required: [seasonId, seasonNumber, teams]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/standings:
    get:
      operationId: getLeagueStandings
      tags: [league]
      summary: Get team standings and season stats
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: Standings with offensive/defensive stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  currentWeek:
                    type: string
                  standings:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamStanding"
                required: [seasonId, seasonNumber, currentWeek, standings]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/schedule:
    get:
      operationId: getLeagueSchedule
      tags: [league]
      summary: Get season schedule
      description: Returns all games for the active season. Filter by week with ?week=N (0-based index)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
        - name: week
          in: query
          description: 0-based week index to filter by
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Schedule games
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  currentWeek:
                    type: string
                  games:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScheduleGame"
                required: [seasonId, seasonNumber, currentWeek, games]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/rosters:
    get:
      operationId: getAllRosters
      tags: [league]
      summary: Get all team rosters
      description: Returns every player from all teams for the active season
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: All rosters
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  rosters:
                    type: array
                    items:
                      $ref: "#/components/schemas/RosterPlayer"
                required: [seasonId, seasonNumber, rosters]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/roster/{teamId}:
    get:
      operationId: getTeamRoster
      tags: [league]
      summary: Get roster for a specific team
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
        - name: teamId
          in: path
          required: true
          description: MCA team ID
          schema:
            type: integer
      responses:
        "200":
          description: Team roster
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  teamId:
                    type: integer
                  roster:
                    type: array
                    items:
                      $ref: "#/components/schemas/RosterPlayer"
                required: [seasonId, seasonNumber, teamId, roster]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/players/stats:
    get:
      operationId: getPlayerStats
      tags: [league]
      summary: Get player season stats
      description: Returns cumulative player stats for the active season. Optionally filter by position.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
        - name: position
          in: query
          required: false
          description: Filter by position abbreviation (e.g. QB, WR, RB)
          schema:
            type: string
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  stats:
                    type: array
                    items:
                      $ref: "#/components/schemas/PlayerSeasonStats"
                required: [seasonId, seasonNumber, stats]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/draft-picks:
    get:
      operationId: getDraftPicks
      tags: [league]
      summary: Get draft pick inventory
      description: Returns all draft picks held by each team for the active season
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: Draft picks
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  picks:
                    type: array
                    items:
                      $ref: "#/components/schemas/DraftPick"
                required: [seasonId, seasonNumber, picks]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/news:
    get:
      operationId: getLeagueNews
      tags: [league]
      summary: Get league news feed
      description: Returns EA CFM news items for the active season
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: League news
          content:
            application/json:
              schema:
                type: object
                properties:
                  seasonId:
                    type: integer
                  seasonNumber:
                    type: integer
                  news:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeagueNewsItem"
                required: [seasonId, seasonNumber, news]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/leagues/{guildId}/users:
    get:
      operationId: getLeagueUsers
      tags: [league]
      summary: Get all league members
      description: Returns all registered users for a guild with balances and all-time records
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/guildId"
      responses:
        "200":
          description: League members
          content:
            application/json:
              schema:
                type: object
                properties:
                  guildId:
                    type: string
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeagueMember"
                required: [guildId, members]
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    guildId:
      name: guildId
      in: path
      required: true
      description: Discord guild (server) ID
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      required: false
      description: Maximum number of results to return (max 200)
      schema:
        type: integer
        default: 50
        maximum: 200
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Season:
      type: object
      properties:
        id:
          type: integer
        guildId:
          type: string
        seasonNumber:
          type: integer
        isActive:
          type: boolean
        currentWeek:
          type: string
        catchupMode:
          type: boolean
        startedAt:
          type: string
          format: date-time
      required: [id, guildId, seasonNumber, isActive, currentWeek, catchupMode, startedAt]
    McaTeam:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        teamId:
          type: integer
        fullName:
          type: string
        nickName:
          type: string
        conference:
          type: string
          nullable: true
        userName:
          type: string
        isHuman:
          type: boolean
        discordId:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
      required: [id, seasonId, teamId, fullName, nickName, userName, isHuman]
    TeamStanding:
      type: object
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        discordId:
          type: string
          nullable: true
        wins:
          type: integer
        losses:
          type: integer
        offYds:
          type: integer
        offPassYds:
          type: integer
        offRushYds:
          type: integer
        offTDs:
          type: integer
        offPtsPerGame:
          type: number
        defPassYds:
          type: integer
        defRushYds:
          type: integer
        defTDs:
          type: integer
        teamSacks:
          type: integer
        teamInts:
          type: integer
        offRedZonePct:
          type: number
        defRedZonePct:
          type: number
        turnoverDiff:
          type: integer
      required: [teamId, teamName, wins, losses]
    ScheduleGame:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        weekIndex:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homeTeamName:
          type: string
        awayTeamName:
          type: string
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        status:
          type: integer
      required: [id, seasonId, weekIndex, homeTeamId, awayTeamId, homeTeamName, awayTeamName, status]
    RosterPlayer:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        discordId:
          type: string
          nullable: true
        playerId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        overall:
          type: integer
        devTrait:
          type: integer
        age:
          type: integer
          nullable: true
        jerseyNum:
          type: integer
          nullable: true
        contractYearsLeft:
          type: integer
          nullable: true
        archetypeAbbrev:
          type: string
          nullable: true
        portraitUrl:
          type: string
          nullable: true
      required: [id, seasonId, teamId, teamName, playerId, firstName, lastName, position, overall, devTrait]
    PlayerSeasonStats:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        playerId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        discordId:
          type: string
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        passYds:
          type: integer
        passTDs:
          type: integer
        passAtt:
          type: integer
        passComp:
          type: integer
        passInts:
          type: integer
        rushYds:
          type: integer
        rushTDs:
          type: integer
        rushAtt:
          type: integer
        recYds:
          type: integer
        recTDs:
          type: integer
        recRec:
          type: integer
        sacks:
          type: number
        defInts:
          type: integer
        totalTackles:
          type: integer
        defTDs:
          type: integer
        fgMade:
          type: integer
        fgAtt:
          type: integer
      required: [id, seasonId, playerId, teamId, teamName, firstName, lastName, position]
    DraftPick:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        teamId:
          type: integer
        teamName:
          type: string
        discordId:
          type: string
          nullable: true
        draftYear:
          type: integer
        round:
          type: integer
        pickNum:
          type: integer
        originalTeamId:
          type: integer
          nullable: true
        originalTeamName:
          type: string
          nullable: true
      required: [id, seasonId, teamId, teamName, draftYear, round, pickNum]
    LeagueNewsItem:
      type: object
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        headline:
          type: string
        body:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        weekIndex:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, seasonId, headline, createdAt]
    LeagueMember:
      type: object
      properties:
        id:
          type: integer
        discordId:
          type: string
        discordUsername:
          type: string
        serverNickname:
          type: string
          nullable: true
        team:
          type: string
          nullable: true
        balance:
          type: integer
        eaId:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        allTimeSuperbowlWins:
          type: integer
        allTimeSuperbowlLosses:
          type: integer
        allTimeH2HWins:
          type: integer
        allTimeH2HLosses:
          type: integer
        playoffSeed:
          type: integer
          nullable: true
        playoffConference:
          type: string
          nullable: true
      required: [id, discordId, discordUsername, balance, isAdmin]
...