# lib/api-zod/src/generated/types/playerSeasonStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerSeasonStats {
  id: number;
  seasonId: number;
  playerId: number;
  teamId: number;
  teamName: string;
  discordId?: string | null;
  firstName: string;
  lastName: string;
  position: string;
  passYds?: number;
  passTDs?: number;
  passAtt?: number;
  passComp?: number;
  passInts?: number;
  rushYds?: number;
  rushTDs?: number;
  rushAtt?: number;
  recYds?: number;
  recTDs?: number;
  recRec?: number;
  sacks?: number;
  defInts?: number;
  totalTackles?: number;
  defTDs?: number;
  fgMade?: number;
  fgAtt?: number;
}