# lib/api-zod/src/generated/types/teamStanding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamStanding {
  teamId: number;
  teamName: string;
  discordId?: string | null;
  wins: number;
  losses: number;
  offYds?: number;
  offPassYds?: number;
  offRushYds?: number;
  offTDs?: number;
  offPtsPerGame?: number;
  defPassYds?: number;
  defRushYds?: number;
  defTDs?: number;
  teamSacks?: number;
  teamInts?: number;
  offRedZonePct?: number;
  defRedZonePct?: number;
  turnoverDiff?: number;
}